Change line-end terminator to UTF-16-CRLF for DSL files.
LingvoDSL defines its format with implicit assumption on
MS Windows platform. That is because it should be CRLF
line end when generating UTF-16LE encoding file.

GoldenDict and other DSL library such as DSL4j try to
interpret various combinations of encoding
and eol terminators, but it is better to use CRLF
to avoid unknown bugs.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>